+++ conrpggame/Adventures/Interfaces/IAdventrueService.cs
@@ -1,0 +1,11 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace conrpggame.Adventures.Interfaces
+{
+    public interface IAdventrueService
+    {
+        Adventrues GetInitalAdventrue();
+    }
+}
